Add support for generating aggregated inlined objects
