+++ src/main/kotlin/com/cjbooms/fabrikt/util/KaizenParserExtensions.kt
@@ -39,10 +39,13 @@ object KaizenParserExtensions {
         getDiscriminatorForInLinedObjectUnderAllOf()?.propertyName != null
 
     fun Schema.isInlinedObjectDefinition() =
-        isObjectType() && !isSchemaLess() && (
+        (isObjectType() || isAggregatedObject()) && !isSchemaLess() && (
             Overlay.of(this).pathFromRoot.contains("properties") ||
                 Overlay.of(this).pathFromRoot.contains("items")
             )
+
+    private fun Schema.isAggregatedObject(): Boolean =
+        allOfSchemas?.isNotEmpty() == true || anyOfSchemas?.isNotEmpty() == true
 
     fun Schema.isInlinedTypedAdditionalProperties() =
         isObjectType() && !isSchemaLess() && Overlay.of(this).pathFromRoot.contains("additionalProperties")
@@ -188,7 +191,6 @@ object KaizenParserExtensions {
     fun Schema.safeName(): String =
         when {
             isOneOfPolymorphicTypes() -> this.oneOfSchemas.first().allOfSchemas.first().safeName()
-            isPropertyWithAllOfSingleType() -> this.allOfSchemas.first().safeName()
             name != null -> name
             else -> Overlay.of(this).pathFromRoot
                 .splitToSequence("/")
@@ -221,26 +223,6 @@ object KaizenParserExtensions {
 
     fun Schema.isOneOfSuperInterface() =
         discriminator != null && discriminator.propertyName != null && oneOfSchemas.isNotEmpty()
-
-    private fun Schema.isPropertyWithAllOfSingleType() =
-        allOfSchemas?.size == 1 && isInlinedPropertySchema()
-
-    /**
-     * The `pathFromRoot` of a property schema ends with
-     * `/properties/<name of property>`, so we check if the
-     * penultimate segment is `properties`.
-     */
-    private fun Schema.isInlinedPropertySchema(): Boolean {
-        val path = Overlay.of(this).pathFromRoot
-        val lastSegment = path.lastIndexOf('/')
-        if (lastSegment != -1) {
-            val penultimateSegment = path.lastIndexOf('/', lastSegment - 1)
-            if (penultimateSegment != -1) {
-                return path.startsWith("/properties/", penultimateSegment)
-            }
-        }
-        return false
-    }
 
     fun OpenApi3.basePath(): String =
         servers
