+++ src/test/kotlin/com/cjbooms/fabrikt/generators/ModelGeneratorTest.kt
@@ -49,7 +49,7 @@ class ModelGeneratorTest {
         "validationAnnotations",
         "wildCardTypes",
         "singleAllOf",
-        "singleAllOfProperty",
+        "inlinedAggregatedObjects",
         "responsesSchema",
         "webhook",
         "instantDateTime",
